Build 0.3

+++ log/models.py
@@ -17,11 +17,11 @@ class Selling_Log(models.Model):
     sell_member = models.CharField(max_length=100)
     sell_type = models.CharField(max_length=1, choices=type)
     sell_employee = models.CharField(max_length=100)
-    cash_money = models.FloatField()
-    transfer_money = models.FloatField()
-    fee = models.FloatField()
-    discount = models.FloatField()
-    total_price = models.FloatField()
+    cash_money = models.DecimalField(max_digits=8, decimal_places=2)
+    transfer_money = models.DecimalField(max_digits=8, decimal_places=2)
+    fee = models.DecimalField(max_digits=8, decimal_places=2)
+    discount = models.DecimalField(max_digits=8, decimal_places=2)
+    total_price = models.DecimalField(max_digits=8, decimal_places=2)
     datetime = models.DateTimeField(auto_now_add=True)
 
     class Meta:
@@ -33,8 +33,8 @@ class Selling_Log(models.Model):
     
 # Selling Detail Log
 class Sell_Detail_Log(models.Model):
-    sell_id  = models.ForeignKey(Selling_Log, to_field='sell_id', on_delete=models.CASCADE)
-    sell_barcode  = models.ForeignKey(System_db.Itemlist, to_field='item_barcode', on_delete=models.CASCADE)
+    sell_id = models.ForeignKey(Selling_Log, to_field='sell_id', on_delete=models.CASCADE)
+    sell_barcode = models.ForeignKey(System_db.Itemlist, to_field='item_barcode', on_delete=models.CASCADE)
     sell_count = models.IntegerField(default=1)
     sell_price = models.FloatField()
